+++ mediation/AppLovin/source/plugin/Assets/GoogleMobileAds/Common/Mediation/AppLovin/IAppLovinClient.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (C) 2017 Google, Inc.
+// Copyright (C) 2017 Google, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -19,16 +19,12 @@ namespace GoogleMobileAds.Common.Mediation.AppLovin
 {
     public interface IAppLovinClient
     {
-        // Initialize the AppLovin SDK.
         void Initialize();
 
-        // Sets a flag indicating whether a user located in the
-        // European Union (i.e., EU/GDPR data subject) has provided
-        // opt-in consent for the collection and use of personal data.
         void SetHasUserConsent(bool hasUserConsent);
 
-        // Sets a flag indicating whether a user is known to be
-        // in an age-restricted category (i.e., under the age of 16).
         void SetIsAgeRestrictedUser(bool isAgeRestrictedUser);
+
+        void SetDoNotSell(bool doNotSell);
     }
 }